Added authentication

+++ backend/controller/todoController.js
@@ -2,7 +2,7 @@ import pool from "../db.js";
 
 const getTodo = async (req,res)=>{
     try{
-    const result = await pool.query("select * from todos");
+    const result = await pool.query("select * from todos where user_id = $1",[req.user]);
     res.json(result.rows);
     }catch(error){
         console.error(error);
@@ -15,13 +15,14 @@ const getTodo = async (req,res)=>{
 const getById =  async (req,res)=>{
     try{
         const id = Number(req.params.id);
+        const user_id = req.user;
         if(!Number.isInteger(id)|| id <=0){
             return res.status(400).json({
             error: "Invalid todo ID"        
         });
         }
-        const text = 'SELECT * FROM todos where id = $1'
-        const value = [id];
+        const text = 'SELECT * FROM todos where id = $1 and user_id = $2'
+        const value = [id,user_id];
         const result = await pool.query(text,value);
     
         if (result.rows.length === 0) {
@@ -43,14 +44,15 @@ const createTodo = async (req,res)=>{
     try{
     const name = req.body.name;
     const priority = req.body.priority;
+    const user_id = req.user;
     if(!name){
         return res.status(400).json({
             error: "name is required"
         })
     }
 
-    const text = 'INSERT INTO todos(name,priority) VALUES ($1, $2) RETURNING *';
-    const values = [name , priority];
+    const text = 'INSERT INTO todos(name,priority,user_id) VALUES ($1, $2,$3) RETURNING *';
+    const values = [name , priority,user_id];
     const result = await pool.query(text,values);
 
     return res.status(201).json(result.rows[0]);
@@ -65,14 +67,15 @@ const createTodo = async (req,res)=>{
 
 const deleteById =  async (req,res)=>{
     try{
+    const user_id = req.user;
     const id = Number(req.params.id);
     if(!Number.isInteger(id)|| id <=0){
         return res.status(400).json({
         error: "Invalid todo ID"        
     });
     }
-    const text = 'DELETE FROM todos WHERE id = $1 RETURNING *';
-    const values = [id];
+    const text = 'DELETE FROM todos WHERE id = $1 AND user_id = $2 RETURNING *';
+    const values = [id,user_id];
     const result = await pool.query(text,values);
       if (result.rows.length === 0) {
     return res.status(404).json({
@@ -92,6 +95,7 @@ const deleteById =  async (req,res)=>{
 const updatebyId = async (req, res) => {
     try{
   const id = Number(req.params.id);
+  const user_id = req.user;
 
   if(!Number.isInteger(id)|| id <=0){
         return res.status(400).json({
@@ -117,11 +121,13 @@ const updatebyId = async (req, res) => {
   }
 
   values.push(id);
+  values.push(user_id)
 
   const text = `
     UPDATE todos
     SET ${setClauses.join(", ")}
-    WHERE id = $${values.length}
+    WHERE id = $${values.length-1} AND
+    user_id =  $${values.length}
     RETURNING *
   `;
